# doc/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "华农掌中行 API",
    "version": "1.0.0",
    "description": "由 back/server.js 与 back/routes/*.js 导出的接口定义（2026-08-14）。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地后端（以 PORT 环境变量为准）"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "用户注册",
        "description": "手机号须为中国大陆手机号；密码至少 6 位",
        "operationId": "post_api_auth_register",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "用户认证"
        ],
        "summary": "用户登录",
        "description": "成功后返回有效期 7 天的 JWT",
        "operationId": "post_api_auth_login",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone",
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/api/auth/userinfo": {
      "get": {
        "tags": [
          "用户认证"
        ],
        "summary": "获取当前用户信息",
        "description": "Authorization 头直接传 token，不加 Bearer 前缀",
        "operationId": "get_api_auth_userinfo",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数或业务状态错误"
          },
          "401": {
            "description": "未认证或 token 无效"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "资源不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "用户 JWT，直接传 token"
      },
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "管理员 JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code"
        ]
      }
    }
  }
}